requirements.txt
boto3==1.34.79
botocore==1.34.79
httpx==0.27.0
requests==2.32.3
streamlit==1.31.1
black==24.3.0
ruff==0.4.1
pytest==8.1.1
fpdf2==2.7.5
pandas==2.2.1
